+++ client/src/components/shared/Carousel.test.jsx
@@ -1,5 +1,5 @@
 import { screen } from '@testing-library/react';
-import Carousel from 'App/Carousel.jsx';
+import Carousel from 'shared/Carousel.jsx';
 
 describe('Carousel', () => {
